requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.17.1
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.4
general_operator==0.3.10
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
influxdb-client==1.44.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
protobuf==5.27.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
reactivex==4.0.4
redis==5.0.7
rich==13.7.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.40.0
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.32.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
